requirements.txt
torch
torchvision
Pillow
torchmetrics